Everything is complitely builded and even works

+++ AbstractShop/AbstractShopView/FormCreateOrder.cs
@@ -89,7 +89,7 @@ namespace LawFirmView
                 _logicO.CreateOrder(new CreateOrderBindingModel
 
                 {                   
-                    Sum = Convert.ToInt32(textBoxSum.Text),
+                    Sum = Convert.ToDecimal(textBoxSum.Text),
                     DocumentId = Convert.ToInt32(comboBoxDocument.SelectedValue),
                     Count = Convert.ToInt32(textBoxCount.Text)
                 }) ;
